[Refactor] Remove unused date time extension and move string null check in primitive extension (#13)
* refactor: Replace string extension imports with primitive types extension and remove string class

* refactor: Remove unused date time extnesion and move string null check in primitive extension

* refactor: Apply requetsed changes from pr review comments
