+++ lib/presentation/login/screens/phone_num_otp_screen/widgets/otp_input_field.dart
@@ -7,7 +7,7 @@ import 'package:skelter/constants/integration_test_keys.dart';
 import 'package:skelter/presentation/login/bloc/login_bloc.dart';
 import 'package:skelter/presentation/login/bloc/login_events.dart';
 import 'package:skelter/presentation/login/bloc/login_state.dart';
-import 'package:skelter/utils/extensions/string.dart';
+import 'package:skelter/utils/extensions/primitive_types_extensions.dart';
 import 'package:skelter/widgets/styling/app_colors.dart';
 import 'package:sms_autofill/sms_autofill.dart';
